refactor activitystreams object registry
