+++ vultron/as_vocab/base/objects/object_types.py
@@ -16,9 +16,9 @@
 from datetime import datetime
 from typing import TypeAlias
 
-from vultron.as_vocab.base import activitystreams_object
 from vultron.as_vocab.base.links import ActivityStreamRef
 from vultron.as_vocab.base.objects.base import as_Object, as_ObjectRef
+from vultron.as_vocab.base.registry import activitystreams_object
 
 
 @activitystreams_object
