+++ vultron/as_vocab/objects/case_status.py
@@ -19,10 +19,10 @@ from typing import TypeAlias
 
 from pydantic import field_serializer, field_validator, model_validator
 
-from vultron.as_vocab.base import activitystreams_object
 from vultron.as_vocab.base.links import as_Link, ActivityStreamRef
 from vultron.as_vocab.base.objects.actors import as_Actor
 from vultron.as_vocab.base.objects.base import as_Object
+from vultron.as_vocab.base.registry import activitystreams_object
 from vultron.as_vocab.objects.base import VultronObject
 from vultron.bt.embargo_management.states import EM
 from vultron.bt.report_management.states import RM
