+++ vultron/as_vocab/objects/vulnerability_case.py
@@ -19,9 +19,9 @@ from typing import TypeAlias
 
 from pydantic import Field, model_validator
 
-from vultron.as_vocab.base import activitystreams_object
 from vultron.as_vocab.base.links import ActivityStreamRef
 from vultron.as_vocab.base.objects.activities.base import as_Activity
+from vultron.as_vocab.base.registry import activitystreams_object
 from vultron.as_vocab.objects.base import VultronObject
 from vultron.as_vocab.objects.case_participant import CaseParticipantRef
 from vultron.as_vocab.objects.case_status import CaseStatusRef, CaseStatus
