+++ vultron/as_vocab/objects/vulnerability_report.py
@@ -17,8 +17,8 @@ Provides a VulnerabilityReport object for the Vultron ActivityStreams Vocabulary
 
 from typing import TypeAlias
 
-from vultron.as_vocab.base import activitystreams_object
 from vultron.as_vocab.base.links import ActivityStreamRef
+from vultron.as_vocab.base.registry import activitystreams_object
 from vultron.as_vocab.objects.base import VultronObject
